Renamed endpoints/usersync to endpoints/cookie_sync, which should be a bit easier to find.
