+++ endpoints/cookie_sync/usersync_test.go
@@ -1,4 +1,4 @@
-package usersync
+package cookie_sync
 
 import (
 	"bytes"
