+++ pbs_light.go
@@ -50,8 +50,8 @@ import (
 	"github.com/prebid/prebid-server/cache/filecache"
 	"github.com/prebid/prebid-server/cache/postgrescache"
 	"github.com/prebid/prebid-server/config"
+	syncEndpoints "github.com/prebid/prebid-server/endpoints/cookie_sync"
 	"github.com/prebid/prebid-server/endpoints/openrtb2"
-	syncEndpoints "github.com/prebid/prebid-server/endpoints/usersync"
 	"github.com/prebid/prebid-server/exchange"
 	"github.com/prebid/prebid-server/openrtb_ext"
 	"github.com/prebid/prebid-server/pbs"
